Start working on the `textif` answers
The visa application has these questions where you select a yes/no
and may be asked a follow-up question. This change starts to deal
with that. We still do not properly set the requirements for the
input and do not hide/display the explanation input automatically,
but hey, it is a start.