Add the visa send-out looper definition
This one is a no-op right now due to being commented out, but should
probably stay this way for going public. It will get an attachment
added with the visa graphic PDF and the whole process will need to be
started off with the user's email being added as a key during the
initial e-mail being sent out.