Add Visa template and name font
The PSD (and therefore PNG as well) are not as high in resolution
as I requested, but will have to do. The font will be required for
entering the person's name on the visa paper form. This is pretty
much an A-series paper size print-out roughly for A5 or A6 printing
with decent-ish quality in the end, which some letterboxing borders.