Sort the XLSX applications by Visa ID
Our Visa IDs can be a good way to sort by.
We will probably have to rip this logic out for the actual list
we provide for border patrol since these IDs are not actually
printed on the final visas being sent out, but this is a nice
thing to do for now and possibly for the next year.